requirements.txt
accelerate==0.26.0
addict==2.4.0
aiofiles==23.2.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
bitsandbytes==0.44.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
deepspeed==0.12.6
diffusers==0.31.0
docker-pycreds==0.4.0
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.2.2
fastapi==0.115.2
ffmpy==0.4.0
filelock==3.16.1
flash-attn==2.6.3
fonttools==4.54.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
gradio==4.16.0
gradio_client==0.8.1
h11==0.14.0
h5py==3.13.0
hjson==3.1.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.30.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
latex2mathml==3.77.0
markdown-it-py==3.0.0
markdown2==2.5.1
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.1
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
orjson==3.10.9
packaging==24.1
pandas==2.2.3
peft==0.13.2
pillow==10.4.0
platformdirs==4.3.6
protobuf==3.19.0
psutil==6.1.0
py-cpuinfo==9.0.0
pyarrow==18.0.0
pycocotools==2.0.8
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pynvml==11.5.3
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
ruff==0.7.0
safetensors==0.4.5
scikit-learn==1.2.2
scipy==1.14.1
seaborn==0.13.2
segment-anything==1.0
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==2.17.0
setproctitle==1.3.3
shellingham==1.5.4
shortuuid==1.0.13
smmap==5.0.1
sniffio==1.3.1
starlette==0.40.0
svgwrite==1.4.3
sympy==1.13.3
threadpoolctl==3.5.0
timm==0.6.13
tokenizers==0.21.0
tomli==2.1.0
tomlkit==0.12.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.5
transformers==4.47.0
triton==2.1.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
wandb==0.18.5

